+++ monitor/ui/config.js
@@ -1,3 +1,3 @@
 window.MONITOR_CONFIG = {
-  apiUrl: 'http://127.0.0.1:4310'
+  apiUrl: 'http://127.0.0.1:4319'
 };